Added Manifest and other improvements
