+++ now.json
@@ -2,7 +2,14 @@
   "version": 2,
   "builds": [
     { "src": "next.config.js", "use": "@now/next" },
-    { "src": "api/**/*", "use": "@now/node" },
+    {
+      "src": "api/**/*",
+      "use": "@now/node",
+      "config": {
+        "includeFiles": ["manifest/**"],
+        "maxLambdaSize": "50mb"
+      }
+    },
     { "src": "static/**/*", "use": "@now/static" }
   ],
   "routes": [
