+++ package.json
@@ -29,6 +29,7 @@
     "react-tooltip": "^3.10.0",
     "redux": "^4.0.1",
     "redux-thunk": "^2.3.0",
+    "sqlite3": "^4.0.8",
     "yarn": "^1.16.0"
   },
   "scripts": {
