+++ src/server/apiConstants.js
@@ -1,0 +1,45 @@
+module.exports = {
+  raceType: {
+    0: "Human",
+    1: "Awoken",
+    2: "Exo",
+    3: "Unknown"
+  },
+  genderType: {
+    0: "Male",
+    1: "Female",
+    2: "Unknown"
+  },
+  classType: {
+    0: "Titan",
+    1: "Hunter",
+    2: "Warlock",
+    3: "Unknown"
+  },
+  itemType: {
+    None: 0,
+    Currency: 1,
+    Armor: 2,
+    Weapon: 3,
+    Message: 7,
+    Engram: 8,
+    Consumable: 9,
+    ExchangeMaterial: 10,
+    MissionReward: 11,
+    QuestStep: 12,
+    QuestStepComplete: 13,
+    Emblem: 14,
+    Quest: 15,
+    Subclass: 16,
+    ClanBanner: 17,
+    Aura: 18,
+    Mod: 19,
+    Dummy: 20,
+    Ship: 21,
+    Vehicle: 22,
+    Emote: 23,
+    Ghost: 24,
+    Package: 25,
+    Bounty: 26
+  }
+};
